+++ magda-web-client/src/Components/SearchFacets/SearchFacets.js
@@ -1,13 +1,19 @@
 import React, { Component } from "react";
 import { config } from "../../config";
+import { Small, Medium } from "../../UI/Responsive";
+import downArrowDark from "../../assets/downArrowDark.svg";
+
 import "./SearchFacets.css";
 
 class SearchFacets extends Component {
     constructor(props) {
         super(props);
-        this.state = { openFacet: null };
+        this.state = { openFacet: null, showFilterOnMobile: false };
         this.toggleFacet = this.toggleFacet.bind(this);
         this.closeFacetWithKeyBoard = this.closeFacetWithKeyBoard.bind(this);
+        this.onToggleFacetOnMobile = this.onToggleFacetOnMobile.bind(this);
+        this.renderMobile = this.renderMobile.bind(this);
+        this.renderDesktop = this.renderDesktop.bind(this);
     }
 
     componentWillMount() {
@@ -51,9 +57,16 @@ class SearchFacets extends Component {
         return false;
     }
 
-    render() {
+    onToggleFacetOnMobile() {
+        this.setState({
+            showFilterOnMobile: !this.state.showFilterOnMobile
+        });
+    }
+
+    renderDesktop() {
         return (
-            <div className="search-facets clearfix">
+            <div className="search-facets-desktop">
+                <div className="sub-heading"> Filters </div>
                 {config.facets.map(c => (
                     <div
                         className="search-facet"
@@ -70,6 +83,48 @@ class SearchFacets extends Component {
                         />
                     </div>
                 ))}
+            </div>
+        );
+    }
+
+    renderMobile() {
+        return (
+            <div className="search-facets-mobile">
+                <button
+                    className="filter-toggle-button mui-btn"
+                    onClick={this.onToggleFacetOnMobile}
+                >
+                    Filters <img src={downArrowDark} alt="open filter" />
+                </button>
+                {this.state.showFilterOnMobile &&
+                    config.facets.map(c => (
+                        <div
+                            className="search-facet"
+                            key={c.id}
+                            onClick={ev => ev.stopPropagation()}
+                        >
+                            <c.component
+                                updateQuery={this.props.updateQuery}
+                                location={this.props.location}
+                                title={c.id}
+                                isOpen={this.state.openFacet === c.id}
+                                toggleFacet={this.toggleFacet.bind(this, c.id)}
+                                closeFacet={this.closeFacet.bind(this, c.id)}
+                            />
+                        </div>
+                    ))}
+                {this.state.openFacet && (
+                    <div className="mobile-facet-background" />
+                )}
+            </div>
+        );
+    }
+
+    render() {
+        return (
+            <div>
+                <Medium>{this.renderDesktop()}</Medium>
+                <Small>{this.renderMobile()}</Small>
             </div>
         );
     }